storage: move code out of region_info_accessor
